refactor: 난수 생성 클래스 구조 변경
