+++ src/main/java/baseball/util/NumberGenerator.java
@@ -1,4 +1,4 @@
-package baseball.model;
+package baseball.util;
 import camp.nextstep.edu.missionutils.Randoms;
 import java.util.ArrayList;
 import java.util.List;
